chore: moved playlistwindow to feature/winamp/playlist
