--- components/ui/Winamp/PlayListWindow/PlayListWindow/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./view";
