+++ src/features/winamp/playlist/constants.ts
@@ -1,0 +1,1 @@
+export const WINDOW_NAME = "PLAYLIST";
